Created MainList component & MainList styled component & refactored code from App components to MainList component
