+++ package.json
@@ -12,8 +12,12 @@
     "react": "^16.13.1",
     "react-beautiful-dnd": "^13.0.0",
     "react-dom": "^16.13.1",
+    "react-redux": "^7.2.1",
     "react-scripts": "3.4.1",
     "react-textarea-autosize": "^8.2.0",
+    "redux": "^4.0.5",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
     "serialize-javascript": "^3.1.0",
     "styled-components": "^5.2.0",
     "websocket-extensions": "^0.1.4"
